Move dev handler.
